Incorporate the r_cmd_check workflow used in the hubverse
To simplify things, let's try using a single workflow for all of the
R checks. This should eliminate the need for a second pr_unittest.yaml
that runs the unit tests and builds the vignettes.

I'm guessing this way will take longer, since it's no longer making some
of the long-running unit tests contingent on specific files being
updated. The tradeoff of using the prescribed, simplified workflow is worthwhile,
and the long-running unit tests are a separate problem to fix in a follow-up.